+++ src/main/kotlin/racing/domain/CarFactory.kt
@@ -1,0 +1,7 @@
+package racing.domain
+
+object CarFactory {
+    fun makeCars(names: Names): Cars {
+        return Cars(names)
+    }
+}
