+++ src/main/kotlin/racing/domain/Movable.kt
@@ -1,4 +1,4 @@
-package racing
+package racing.domain
 
 fun interface Movable {
     fun canMove(): Boolean
