+++ src/main/kotlin/racing/domain/RandomMoving.kt
@@ -1,4 +1,4 @@
-package racing
+package racing.domain
 
 object RandomMoving : Movable {
     private const val MIN_MOVABLE_VALUE = 4
