+++ src/main/kotlin/racing/domain/vo/CarInput.kt
@@ -1,3 +1,5 @@
-package racing.model
+package racing.domain.vo
+
+import racing.domain.Names
 
 data class CarInput(val names: Names, val times: Times)
