+++ src/main/kotlin/racing/domain/vo/CarMoveResult.kt
@@ -1,3 +1,3 @@
-package racing.model
+package racing.domain.vo
 
 data class CarMoveResult(val name: Name, val position: Position)
