--- src/main/kotlin/racing/model/CarFactory.kt
@@ -1,7 +1,0 @@
-package racing.model
-
-object CarFactory {
-    fun makeCars(names: Names): Cars {
-        return Cars(names.names.map { Car(it, Position()) })
-    }
-}
