+++ src/test/kotlin/racing/PositionTest.kt
@@ -4,7 +4,7 @@ import org.junit.jupiter.api.assertThrows
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ValueSource
 import racing.exception.PositionException
-import racing.model.Position
+import racing.domain.vo.Position
 
 class PositionTest {
     @ParameterizedTest
